Minor fix on logic, add logic for sounds/notifications

+++ PomoApp.Core/Settings.cs
@@ -2,8 +2,8 @@ namespace PomoApp.Core;
 
 public class Settings
 {
-    public int WorkMinutes { get; set; } = 25;
-    public int BreakMinutes { get; set; } = 5;
-    public int LongBreakMinutes { get; set; } = 15;
+    public double WorkMinutes { get; set; } = 25;
+    public double BreakMinutes { get; set; } = 5;
+    public double LongBreakMinutes { get; set; } = 15;
     public int SessionsBeforeLongBreak { get; set; } = 4;
 }
